+++ test/chez/deps-alias/app/dev/devmain.clj
@@ -1,0 +1,2 @@
+(ns devmain)
+(defn -main [& args] (println "devmain"))