+++ test/chez/deps-alias/app/src/appc.clj
@@ -1,0 +1,2 @@
+(ns appc (:require [libc.core :as c]))
+(defn -main [& args] (println "libc" c/version))