+++ test/chez/deps-alias/app/src/appmain.clj
@@ -1,0 +1,2 @@
+(ns appmain)
+(defn -main [& args] (println "main1"))