+++ test/chez/deps-alias/app/src/appmain2.clj
@@ -1,0 +1,2 @@
+(ns appmain2)
+(defn -main [& args] (println "main2"))