+++ test/chez/deps-alias/app/src/appver.clj
@@ -1,0 +1,2 @@
+(ns appver (:require [liba.core :as a]))
+(defn -main [& args] (println "liba" a/version))