+++ test/chez/deps-alias/app/src/appzone.clj
@@ -1,0 +1,2 @@
+(ns appzone)
+(defn -main [& args] (println (java.time.ZoneId/of "UTC")))