+++ test/chez/deps-alias/liba/src/liba/core.clj
@@ -1,0 +1,2 @@
+(ns liba.core)
+(def version "A")