+++ test/chez/deps-alias/libb/src/liba/core.clj
@@ -1,0 +1,2 @@
+(ns liba.core)
+(def version "B")