+++ test/chez/deps-alias/libc/src/libc/core.clj
@@ -1,0 +1,2 @@
+(ns libc.core)
+(def version "C")